feat(staking module): add slashing implementation
